Minor fix to an issue that was causing failure when string input was given for latitude and liongitude.

+++ neon-server/src/main/groovy/com/ncc/neon/query/transform/GeoGridTransformer.groovy
@@ -45,8 +45,8 @@ class GeoGridTransformer implements Transformer {
 
 		queryResult.getData().each { point ->
 			def pointAgg = getFieldValue(point, params.aggregationField)
-			int horizBox = (((getFieldValue(point, params.lonField) as int) - params.minLon) / boxWidth) as int
-			int vertBox = ((params.maxLat - (getFieldValue(point, params.latField) as int)) / boxHeight) as int
+			int horizBox = (((getFieldValue(point, params.lonField) as double) - params.minLon) / boxWidth) as int
+			int vertBox = ((params.maxLat - (getFieldValue(point, params.latField) as double)) / boxHeight) as int
 			if(horizBox >= numHorizTiles || horizBox < 0 || vertBox >= numVertTiles || vertBox < 0) {
 				return
 			}
